Handle 401 error in edit-question-context

+++ src/components/edit-question/choose-question-type.component.jsx
@@ -102,9 +102,10 @@ const ChooseQuestionType = ({ onDeleteQuestion, readOnly = false }) => {
       break;
   }
 
-  const handleDelete = () => {
-    deleteQuestion();
-    onDeleteQuestion();
+  const handleDelete = async () => {
+    if (await deleteQuestion()) {
+      onDeleteQuestion();
+    }
   };
 
   return (
